[Chore #163114505] Rename validator
